각 사용자의 문서에 답변 저장
