+++ Screens/Questions.js
@@ -1,18 +1,34 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import {View, StyleSheet} from 'react-native';
 import  {db}  from '../firebaseConfig';
-import { doc,updateDoc, } from "firebase/firestore"; 
+import { doc,updateDoc, getDoc} from "firebase/firestore"; 
 
 import SetQuestion from '../Components/SetQuestion';
 
 const Questions = (userInfo) => {
+  const[Questions, setQuestions] = useState({});
+  const[selected, setSelected] = useState([]);
+  const[filled, setFilled] = useState([]);
+  const[temp, setTemp] = useState("");
 
   // Submit 버튼을 누를 때 마다 수정된 사항이 적용됨.
   useEffect(() => {
+    console.log(userDepartment)
     updateDB(userId);
+    readfromDB();
   }, []);
 
-  const updateDB = async (id)=>{
+  const readfromDB = async () =>{
+    try{
+        const data = await getDoc(doc(db, "Question" ,userDepartment))
+        setQuestions(data.data())
+
+    }catch(error){
+        console.log(error.message)
+    }
+}
+
+  const updateDB = async (id, filled, selected)=>{
     try{
       const docRef = doc(db, "User", id);
       await updateDoc(docRef, {
@@ -22,6 +38,18 @@ const Questions = (userInfo) => {
         Weight: userWeight,
         Gender: userGender,
         Department: userDepartment,
+      });
+    }catch(error){
+      console.log(error.message)
+    }
+  }
+
+  const updateAnswerDB = async (filled, selected) => {
+    try{
+      const docRef = doc(db, "User", userId);
+      await updateDoc(docRef, {
+        Filled: filled,
+        Selected: selected,
       });
     }catch(error){
       console.log(error.message)
@@ -45,7 +73,7 @@ const Questions = (userInfo) => {
       <Text>User Weight : {JSON.stringify(userWeight)}</Text>
       <Text>User Gender : {JSON.stringify(userGender)}</Text>
       <Text>User Gender : {JSON.stringify(userDepartment)}</Text> */}
-      <SetQuestion userDepartment={userDepartment}/>
+      <SetQuestion  Questions={Questions} onUpdateAnswerDB={updateAnswerDB} selected={selected} setSelected={setSelected} filled={filled} setFilled={setFilled} temp={temp} setTemp={setTemp}/>
     </View>       
   );
 }
